Do not emit mutations if signed zero float/double
There are two representations of zero in IEEE 754. The negative and the
positive zero. This commutis removes a mutation that causes the sign to be flipped.

These mutations are very uneasy to kill. Currently only these side
effects are known:

1 / -0.0 => -Infinity
1 / 0.0 => Infinity
0.0.to_s => "0.0"
-0.0.to_s => "-0.0"

I'm happy to readd the mutation - when someone adds more wisdom to
this case ;)

Closes #126