Use a more error tolerant way to access {Module,Class}#name
"Many" library authors override Class or Modules #name not to return a
String or nil. They are not aware these are integral APIs for
reflection / tools. This change prints an ugly warning message in case
the API was violated.

Mutant used to crash often because of this LSP violation. This change
should provide a generic workaround.