Bubble exceptions in isolation up to parent process
If an exception is raised during the setup or tear down for the mutant it will cause the block to return no data. This causes Marshal.load to fail with a vague error message: "marshal data too short".

Instead, the block should catch the exception and return a stringified version. Then the isolation can raise an exception with the string.