Add "expected_errors" to corpus integrations
This change drops the concept of `excludes` from the corpus integration
configuration. Instead with `expected_errors` you specify errors you
expect to be raised by certain files. The corpus will fail if an
expected error is not encountered when mutating a file, or if an error
is encountered while mutating a file that was not whitelisted