Fix indeterminism from include_context condition
It looks like this guard clause was causing issues that led to
the build sometimes failing with

    super: no superclass method `expression'

this change fixes this issue.

This change also locks devtools to 0.1.10 and parser to 2.3.1.2.
New reek smells have been released which need to be
addressed in a followup. The newer parser version also requires
an update to the warnings file.