Remove warning capture
The warning capture used to be used by mutant in various scenarios
where the act of inserting the mutation (or preparing the environment
in case of zombification) intentionally yields warnings.

Also it used to be used for private extensions to allow to cover
mutations on newly emitted warnings.

While all these goals above are important: They rely on a stable warning
capture mechanism. The mechanism present was not very stable. See the
removed verbosed docs in `Mutant::Warnings`.

The problem is that the core interface for capturing warnings in Ruby is
broken. It invites people to overwrite a *unique global resource* via a
monkeypatch. Various players interact with this method, and mutant also
doing so will cause mutants presence to be measured from specs where it
should not.

For now the added noise from leaking some mutation testing specific
warnings is less a problem than mutants inference with the warning
mechanism.