Remove unconditional mutation to super()
This mutation only makes sense if the method was called with arguments in the first place.
If there are no arguments `super()` and `super` are semantically equivalent.
And there is no vector towards reduced semantics, which is required for an operator to be valid.

If mutant would be able to emit this mutation conditionally (on the presence of arguments) than this is
a valid operator. Until than we cannot emit it.

[fix #615]